Move generic relation fields to generics app
