+++ EquiTrack/EquiTrack/settings/local.py
@@ -55,7 +55,6 @@ if 'test' in sys.argv:
     TEST_NON_SERIALIZED_APPS = [
         # These apps contains test models that haven't been created by migration.
         # So on the serialization stage these models do not exist.
-        'utils.common',
         'rest_extra',
     ]
